fix/#9 : 변수 수정
