+++ app/src/main/java/com/sopt/now/data/remote/dto/response/reqres/RequreUserList.kt
@@ -1,5 +1,6 @@
 package com.sopt.now.data.remote.dto.response.reqres
 
+import com.sopt.now.model.Profile
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
@@ -7,11 +8,42 @@ import kotlinx.serialization.Serializable
 data class RequreUserList(
     @SerialName("data")
     val data: List<Data>,
+    @SerialName("page")
     val page: Int,
     @SerialName("per_page")
-    val per_page: Int,
+    val perPage: Int,
+    @SerialName("support")
     val support: Support,
+    @SerialName("total")
     val total: Int,
     @SerialName("total_pages")
-    val total_pages: Int,
+    val totalPages: Int,
+) {
+    @Serializable
+    data class Data(
+        @SerialName("avatar")
+        val avatar: String,
+        @SerialName("email")
+        val email: String,
+        @SerialName("first_name")
+        val firstName: String,
+        @SerialName("id")
+        val id: Int,
+        @SerialName("last_name")
+        val lastName: String,
+    )
+
+    @Serializable
+    data class Support(
+        @SerialName("text")
+        val text: String,
+        @SerialName("url")
+        val url: String,
+    )
+}
+
+fun RequreUserList.Data.toFriend() = Profile.FriendProfile(
+    profileImage = avatar,
+    name = firstName + lastName,
+    selfDescription = email
 )
