--- app/src/main/java/com/sopt/now/data/remote/dto/response/reqres/Support.kt
@@ -1,9 +1,0 @@
-package com.sopt.now.data.remote.dto.response.reqres
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Support(
-    val text: String,
-    val url: String,
-)
